runfile_path in :bootstrap, agent_options param
The runfile_path parameter is now valid in :bootstrap mode as well.
If present, the given file will be executed explicitely with cf-agent
after the bootstrap (and initial run, if it's a policy hub)

Added new agent_options parameter, which can be used to specify
additional arbitrary options to pass to cf-agent. These options
will only be used in the "main" calls to cf-agent, i.e. not in
the bootstrap or failsafe calls.